Restructure z3fd
Types from pychunked_data_view are now re-exported by z3fdb. All unused
code has been removed. And public / private parts have been separated.
