+++ src/z3fdb/__init__.py
@@ -1,0 +1,32 @@
+# (C) Copyright 2025- ECMWF.
+#
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
+from pychunked_data_view.chunked_data_view import (
+    AxisDefinition,
+    ChunkedDataViewBuilder,
+    ExtractorType,
+)
+from z3fdb.z3fdb import (
+    FdbSource,
+    FdbZarrArray,
+    FdbZarrGroup,
+    FdbZarrStore,
+)
+
+from z3fdb.z3fdb_error import Z3fdbError
+
+__all__ = [
+    "AxisDefinition",
+    "ChunkedDataViewBuilder",
+    "ExtractorType",
+    "FdbSource",
+    "FdbZarrArray",
+    "FdbZarrGroup",
+    "FdbZarrStore",
+    "Z3fdbError",
+]
